requirements.txt

fastapi==0.101.0
uvicorn[standard]==0.22.0
streamlit==1.25.0
requests==2.31.0


numpy==1.27.4
pandas==2.1.1
tqdm==4.66.1


python_docx


sentence-transformers==2.2.2
torch==2.1.0
transformers==4.34.0


faiss-cpu==1.7.4


